Move yaml-cfg to separerate file and add directory support
